Version 2.2.0.8 Beta 1 ================= bear: - Added two new filters based on the TSFileSource & TSFileSink called TSParserSource & TSParserSink. - Implemented a SharedMemory class to mimic the Win32 API File Handler methods in shared memory. - Added playing of Shared Memory files to the TSFileSource & TSParserSource Filters - Added a input pin with its own MultiMemWriter class so it can timeshift to shared memory of incomming streams. - Fixed various bug in the filters for playing transport stream files. - Added reference clock speed change support to the Source Filter's property pages for testing.
git-svn-id: http://svn.nate.deepcreek.org.au/svn/TSFileSource/trunk@124 89fdcf3f-49f1-7541-bbf8-3596d8354d9d

+++ src/LogFileWriter.cpp
@@ -86,7 +86,7 @@ BOOL LogFileWriter::Open(LPCWSTR filename, BOOL append)
 		strCopy(pPartial, filename, pBackslash - filename);
 
 		HANDLE hDir;
-		hDir = CreateFile(W2T(pPartial), GENERIC_READ, FILE_SHARE_READ|FILE_SHARE_DELETE, NULL, OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS, NULL);
+		hDir = CreateFile(W2T(pPartial), (DWORD) GENERIC_READ, (DWORD) (FILE_SHARE_READ|FILE_SHARE_DELETE), NULL, (DWORD) OPEN_EXISTING, (DWORD) FILE_FLAG_BACKUP_SEMANTICS, NULL);
 		if (hDir == INVALID_HANDLE_VALUE)
 		{
 			if (!CreateDirectory(W2T(pPartial), NULL))
@@ -94,7 +94,7 @@ BOOL LogFileWriter::Open(LPCWSTR filename, BOOL append)
 				delete[] pPartial;
 				return FALSE;
 			}
-			hDir = CreateFile(W2T(pPartial), GENERIC_READ, FILE_SHARE_READ|FILE_SHARE_DELETE, NULL, OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS, NULL);
+			hDir = CreateFile(W2T(pPartial), (DWORD) GENERIC_READ, (DWORD) (FILE_SHARE_READ|FILE_SHARE_DELETE), NULL, (DWORD) OPEN_EXISTING, (DWORD) FILE_FLAG_BACKUP_SEMANTICS, NULL);
 			if (hDir == INVALID_HANDLE_VALUE)
 			{
 				delete[] pPartial;
@@ -111,14 +111,14 @@ BOOL LogFileWriter::Open(LPCWSTR filename, BOOL append)
 	//Open the file for writing
 	if (append)
 	{
-		m_hFile = CreateFile(W2T(filename), GENERIC_WRITE, NULL, NULL, OPEN_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL);
+		m_hFile = CreateFile(W2T(filename), (DWORD) GENERIC_WRITE, (DWORD) NULL, NULL, (DWORD) OPEN_ALWAYS, (DWORD) FILE_ATTRIBUTE_NORMAL, NULL);
 		LARGE_INTEGER li;
 		li.QuadPart = 0;
 		::SetFilePointer(m_hFile, li.LowPart, &li.HighPart, FILE_END);
 	}
 	else
 	{
-		m_hFile = CreateFile(W2T(filename), GENERIC_WRITE, NULL, NULL, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL);
+		m_hFile = CreateFile(W2T(filename), (DWORD) GENERIC_WRITE, (DWORD) NULL, NULL, (DWORD) CREATE_ALWAYS, (DWORD) FILE_ATTRIBUTE_NORMAL, NULL);
 	}
 	return (m_hFile != INVALID_HANDLE_VALUE);
 }
